+++ lib/chromic_pdf/pdf/protocols/close_target.ex
@@ -1,0 +1,11 @@
+defmodule ChromicPDF.CloseTarget do
+  @moduledoc false
+
+  import ChromicPDF.ProtocolMacros
+
+  steps do
+    call(:close_target, "Target.closeTarget", [:targetId], %{})
+    await_response(:target_closed, ["success"])
+    reply("success")
+  end
+end