+++ lib/chromic_pdf/pdf/protocols/spawn_session.ex
@@ -6,18 +6,12 @@ defmodule ChromicPDF.SpawnSession do
   @version Mix.Project.config()[:version]
 
   steps do
-    call(:create_browser_context, "Target.createBrowserContext", [], %{"disposeOnDetach" => true})
-    await_response(:browser_context_created, ["browserContextId"])
-
-    call(:create_target, "Target.createTarget", ["browserContextId"], %{"url" => "about:blank"})
-    await_response(:target_created, ["targetId"])
-
-    call(:attach, "Target.attachToTarget", ["targetId"], %{"flatten" => true})
+    call(:attach, "Target.attachToTarget", [:targetId], %{"flatten" => true})
 
     await_notification(
       :attached,
       "Target.attachedToTarget",
-      [{["targetInfo", "targetId"], "targetId"}],
+      [{["targetInfo", "targetId"], :targetId}],
       ["sessionId"]
     )